+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/debug/TwoDiagnosticReportsTogetherTest.groovy
@@ -1,0 +1,42 @@
+package ru.vyarus.dropwizard.guice.config.debug
+
+import io.dropwizard.Application
+import io.dropwizard.Configuration
+import io.dropwizard.setup.Bootstrap
+import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
+import spock.lang.Specification
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 12.08.2019
+ */
+@UseGuiceyApp(AIApp)
+class TwoDiagnosticReportsTogetherTest extends Specification {
+
+    def "Check diagnostic reports together"() {
+
+        // actual reporting checked manually (test used for reporting configuration)
+
+        expect: "checks that reporting doesn't fail"
+        true
+
+    }
+
+    static class AIApp extends Application<Configuration> {
+
+        @Override
+        void initialize(Bootstrap<Configuration> bootstrap) {
+            bootstrap.addBundle(GuiceBundle.builder()
+                    .printAvailableInstallers()
+                    .printDiagnosticInfo()
+                    .build())
+        }
+
+        @Override
+        void run(Configuration configuration, Environment environment) throws Exception {
+
+        }
+    }
+}